modified CountryResponse

+++ src/main/java/com/example/demo/api/country/dtos/CountryResponse.java
@@ -2,11 +2,13 @@ package com.example.demo.api.country.dtos;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record CountryResponse(
         Integer id,
         String name,
-        Integer area,
+        BigDecimal area,
         String countryCode2
 ) {
 }
